Rename destination container to proxy-api
