+++ cli/cmd/testdata/install_single_namespace_output.golden
@@ -172,7 +172,7 @@ spec:
             port: 9995
         resources: {}
       - args:
-        - destination
+        - proxy-api
         - -addr=:123
         - -controller-namespace=Namespace
         - -single-namespace=true
@@ -185,7 +185,7 @@ spec:
             path: /ping
             port: 9999
           initialDelaySeconds: 10
-        name: destination
+        name: proxy-api
         ports:
         - containerPort: 123
           name: grpc
