+++ controller/api/proxy/endpoint_listener.go
@@ -1,4 +1,4 @@
-package destination
+package proxy
 
 import (
 	"fmt"
