+++ controller/api/proxy/k8s_resolver_test.go
@@ -1,4 +1,4 @@
-package destination
+package proxy
 
 import (
 	"fmt"
