+++ controller/api/proxy/profile_watcher.go
@@ -1,4 +1,4 @@
-package destination
+package proxy
 
 import (
 	"fmt"
