+++ controller/api/proxy/profile_watcher_test.go
@@ -1,4 +1,4 @@
-package destination
+package proxy
 
 import (
 	"reflect"
