+++ controller/api/proxy/server_test.go
@@ -1,4 +1,4 @@
-package destination
+package proxy
 
 import (
 	"context"
