+++ controller/api/proxy/test_helper.go
@@ -1,4 +1,4 @@
-package destination
+package proxy
 
 import (
 	"context"
